requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
langgraph>=0.2.40,<0.3
langchain-core>=0.3.15,<0.3.30
langchain-groq>=0.2.0,<0.3
pydantic>=2.5,<3.0
pydantic-settings>=2.0,<3.0
python-dotenv>=1.0,<2.0
psycopg[binary]>=3.1,<4.0
sqlalchemy>=2.0,<3.0
